Refactor: Split habit feature components
