+++ src/pages/HomePage.tsx
@@ -1,5 +1,5 @@
-import HabitsList from "@/features/habits/components/HabitsList.tsx";
-import Header from "@/features/habits/components/Header.tsx";
+import HabitsList from "@/features/habits/components/habitsList/HabitsList.tsx";
+import Header from "@/features/habits/components/header/Header.tsx";
 
 const HomePage = () => {
   return (
